+++ src/JsonApiSteps.php
@@ -5,13 +5,15 @@ namespace Exozet\Behat\Utils\Base;
 use Behat\Gherkin\Node\TableNode;
 use Behat\Gherkin\Node\PyStringNode;
 
-trait JsonApiSteps {
+trait JsonApiSteps
+{
 
     /**
      * @Then /^the attribute "([^"]*)" has the value "([^"]*)"$/
      * @Then /^Attribut "([^"]*)" hat den Wert "([^"]*)"$/
      */
-    public function theAttributeKeyHasTheValue($propertyName, $expectedValue) {
+    public function theAttributeKeyHasTheValue($propertyName, $expectedValue)
+    {
         $currentValue = json_decode($this->getSession()->getDriver()->getContent(), true);
         $expectedValues = array();
         $expectedValues[$propertyName] = $expectedValue;
@@ -23,8 +25,8 @@ trait JsonApiSteps {
      * @Then /^the object "([^"]*)" contains:$/
      * @Then /^das Objekt "([^"]*)" beinhaltet:$/
      */
-    public function objectKeyContains($propertyName, TableNode $valueTable = null, PyStringNode $pyStringNode = null){
-
+    public function objectKeyContains($propertyName, TableNode $valueTable = null, PyStringNode $pyStringNode = null)
+    {
         $currentValue = $this->getJsonApiValue($propertyName);
         $expectedValues = null;
         if ($pyStringNode != null) {
@@ -109,11 +111,11 @@ trait JsonApiSteps {
          * FIXME: BrowserKitDriver says serverParameters is PRIVATE for whatever reasons. Thus we have no way but
          * reflection to access this value.
          */
-        $class = new \ReflectionClass("Behat\Mink\Driver\BrowserKitDriver");
-        $property = $class->getProperty("serverParameters");
+        $class = new \ReflectionClass('Behat\Mink\Driver\BrowserKitDriver');
+        $property = $class->getProperty('serverParameters');
         $property->setAccessible(true);
         $server = $property->getValue($driver);
-        $server['CONTENT_TYPE'] = "application/json";
+        $server['CONTENT_TYPE'] = 'application/json';
 
         $content = json_encode($values);
         $changeHistory = true;
@@ -155,11 +157,11 @@ trait JsonApiSteps {
          * FIXME: BrowserKitDriver says serverParameters is PRIVATE for whatever reasons. Thus we have no way but
          * reflection to access this value.
          */
-        $class = new \ReflectionClass("Behat\Mink\Driver\BrowserKitDriver");
-        $property = $class->getProperty("serverParameters");
+        $class = new \ReflectionClass('Behat\Mink\Driver\BrowserKitDriver');
+        $property = $class->getProperty('serverParameters');
         $property->setAccessible(true);
         $server = $property->getValue($driver);
-        $server['CONTENT_TYPE'] = "application/x-www-form-urlencoded";
+        $server['CONTENT_TYPE'] = 'application/x-www-form-urlencoded';
 
         $content = http_build_query($values);
         $changeHistory = true;
@@ -176,13 +178,14 @@ trait JsonApiSteps {
      * @param $propertyName
      * @return mixed
      */
-    protected function getJsonApiValue($propertyName) {
+    protected function getJsonApiValue($propertyName)
+    {
         $value = @json_decode($this->getSession()->getDriver()->getContent(), true);
         if ($value === null) {
-            throw new \InvalidArgumentException("The response is not valid JSON!");
+            throw new \InvalidArgumentException('The response is not valid JSON!');
         }
         if (!array_key_exists($propertyName, $value)) {
-            throw new \InvalidArgumentException("The response does not contain a '" . $propertyName . "' key.");
+            throw new \InvalidArgumentException('The response does not contain a "' . $propertyName . '" key.');
         }
         return $value[$propertyName];
     }
@@ -194,13 +197,14 @@ trait JsonApiSteps {
      * @param $path
      * @throws \Exception
      */
-    protected function assertDeepArray($actual, $expected, $path) {
+    protected function assertDeepArray($actual, $expected, $path)
+    {
         if (is_array($expected)) {
             foreach ($expected as $key => $value) {
                 $subPath = $path;
                 $subPath[] = $key;
                 if (!array_key_exists($key, $actual)) {
-                    throw new \Exception("At " . implode('.', $subPath) . " we expect any value, but the key is not set. Expected Key: '" . $key . "'");
+                    throw new \Exception('At ' . implode('.', $subPath) . ' we expect any value, but the key is not set. Expected Key: "' . $key . '"');
                 }
 
 
@@ -211,34 +215,34 @@ trait JsonApiSteps {
             $actualDate = null;
             /* handle 2015-12-17T10:46:00+01:00 */
             if (!is_array($actual)) {
-                $actualDate = \DateTime::createFromFormat("Y-m-d\\TH:i:sO", $actual);
+                $actualDate = \DateTime::createFromFormat('Y-m-d\\TH:i:sO', $actual);
                 if (!$actualDate) {
                     /* handle 2015-12-17T10:46:00.000+01:00 */
-                    $actualDate = \DateTime::createFromFormat("Y-m-d\\TH:i:s.uO", $actual);
+                    $actualDate = \DateTime::createFromFormat('Y-m-d\\TH:i:s.uO', $actual);
                 }
             }
 
             $expectedDate = null;
             /* handle 2015-12-17T10:46:00+01:00 */
             if (!is_array($expected)) {
-                $expectedDate = \DateTime::createFromFormat("Y-m-d\\TH:i:sO", $expected);
+                $expectedDate = \DateTime::createFromFormat('Y-m-d\\TH:i:sO', $expected);
                 if (!$expectedDate) {
                     /* handle 2015-12-17T10:46:00.000+01:00 */
-                    $expectedDate = \DateTime::createFromFormat("Y-m-d\\TH:i:s.uO", $expected);
+                    $expectedDate = \DateTime::createFromFormat('Y-m-d\\TH:i:s.uO', $expected);
                 }
             }
 
             if ($expected !== '*' && $actualDate && $expectedDate) {
                 /* handle two dates */
-                $actualDate->setTimezone(new \DateTimeZone("UTC"));
-                $expectedDate->setTimezone(new \DateTimeZone("UTC"));
+                $actualDate->setTimezone(new \DateTimeZone('UTC'));
+                $expectedDate->setTimezone(new \DateTimeZone('UTC'));
 
                 if ($actualDate->format(\DateTime::RFC3339) != $expectedDate->format(\DateTime::RFC3339)) {
-                    throw new \Exception("At " . implode('.', $path) . " we have the date " . var_export($actual, true) . ", but we expect " . var_export($expected, true) . "");
+                    throw new \Exception('At ' . implode('.', $path) . ' we have the date ' . var_export($actual, true) . ', but we expect ' . var_export($expected, true));
                 }
             } elseif ($actual !== $expected && $expected !== '*') {
                 /* everything except dates */
-                throw new \Exception("At " . implode('.', $path) . " we have the value " . var_export($actual, true) . ", but we expect " . var_export($expected, true) . "");
+                throw new \Exception('At ' . implode('.', $path) . ' we have the value ' . var_export($actual, true) . ', but we expect ' . var_export($expected, true));
             }
         }
     }